1421 march second Bri and Taiga Functional
Modified for two Ps4 controllers at 20% arm motor speeds. This worked and the robot was finally able to drive reasonably well for a couple matches. 
Of note, the robot did reset from a brownout in competition while in a pushing match. Code could be added to reduce power to the Raptor tracks as voltage drops to prevent this.