Last Day of Comp 2023 March 3rd
This Code ended up not being used in competition becasue we were not selected for an alliance. However, the Auto code finally seemed to work. It would drive to 20 inches, then 40 inches, then 60 inches in the poorly named "drive 12 feet" Auto code.
This would not work when selecting Auto a second time, or after having entered Tele-op First.